abstract dissect plugin vs. pprint application
