+++ .gitignore
@@ -4,4 +4,4 @@
 Session.vim
 /test
 /crc
-/dissect
+/pprint
